Fix minimality parameter uniqueness bug
By switching from parameter references to names, the parameters are no
longer unique to each message.

Before, if 'id' was in both messages A and B, it would generate the
clause And(Impl(id, A), Impl(id, B)), which is clearly wrong.

Now it correctly generates And(Impl(id, Or(A, B))).